+++ src/eventsub/channel/subscription/mod.rs
@@ -6,8 +6,11 @@ use serde::{Deserialize, Serialize};
 
 pub mod end;
 pub mod gift;
+pub mod message;
 
 #[doc(inline)]
 pub use end::{ChannelSubscriptionEndV1, ChannelSubscriptionEndV1Payload};
 #[doc(inline)]
 pub use gift::{ChannelSubscriptionGiftV1, ChannelSubscriptionGiftV1Payload};
+#[doc(inline)]
+pub use message::{ChannelSubscriptionMessageV1, ChannelSubscriptionMessageV1Payload};